requirements.txt
fastapi==0.78.0
openai==1.6.1
python-dotenv==0.20.0
pydantic==1.10.11
langchain==0.0.352
beautifulsoup4==4.11.1
tiktoken==0.5.2
uvicorn==0.18.2
googlemaps==4.10.0
tripadvisorapi==1.0.0